(untested) Try to fix config.w32: change build dir to subdir, check for apcu
Do the same thing as solr does: Change the build directory to src/php5
or src/php7, depending on the installed DLL version
(There might be a different way based on php_version.h, not sure)

Try to fix APCU support for igbinary while I'm at it.

- config.w32 wasn't updated to check for apcu. Do that.

Tests are waiting for a windows VM to be set up.