Make array serialization more efficient in PHP 7.
This deduplicates arrays in the serialization if they point
to the same location in memory.
(This does not deduplicate arrays which are equivalent but have
different locations in memory).
PHP 5 is unaffected.

Instead of using the zval pointer of the IS_ARR zval, which would always
be different, use the zend_array storage that it is pointing to.

This assumes that if there are references, they point to different
arrays.

Avoid using ZVAL_EMPTY_ARRAY for now when unserializing,
this would segfault with a naive attempt to use it without extra checks
if the array is immutable (e.g. including the reference count).

- This is less necessary once the changes to aggressively deduplicate
  arrays when serializing are applied
- Unserialization of arrays that are duplicated in many places
  should be faster and use up less memory
  if the new serialization is used.

  The serialization may vary (but be an equivalent representation)
  based on php minor versions and with opcache settings.

Add tests that empty arrays continue to work as expected
(useful to avoid regressions if this uses ZVAL_EMPTY_ARRAY again)

Other changes
- Add miscellaneous micro-optimizations

Related to #182 - Using ZVAL_EMPTY_ARRAY directly when unserializing
would just cause a segfault when combined with references,
without further work.