3.1.6RC1: Fix igbinary_serialize incorrectly deduplicating arrays/objects
... That are freed

For example, this can be seen when `__serialize` returns an array that
contains a temporary array.
Before, because the result of `__serialize` was immediately freed,
another array could be created with the exact same memory address,
and igbinary would incorrectly assume they're duplicates.

Incorrect deduplication can be avoided by keeping references to all
serialized arrays/objects/references around in memory.

This may affect the performance of serialization.

Fixes #287
Based on #288